+++ include/utils/File.h
@@ -8,7 +8,6 @@
 #ifndef _UTILS_FILE_H_
 #define _UTILS_FILE_H_
 
-#include <memory>
 #include <string>
 
 class Folder;
@@ -16,14 +15,13 @@ class Folder;
 class File {
 public:
     File(std::string file);
-    ~File();
 
-    std::shared_ptr<Folder> getParent();
+    std::string &getName();
+    std::string &getPath();
 
 private:
     std::string name;
     std::string path;
-    std::weak_ptr<Folder> parent;
 };
 
 #endif